requirements.txt
asgiref==3.11.1
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.2.11
django-environ==0.12.0
django-import-export==4.4.0
et_xmlfile==2.0.0
mysqlclient==2.2.8
numpy==2.2.6
odfpy==1.4.1
openpyxl==3.1.5
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
sqlparse==0.5.5
tablib==3.9.0
tabulate==0.9.0
typing_extensions==4.15.0
tzdata==2025.3
xlrd==2.0.2
xlwt==1.3.0
